+++ bc/navigation/models.py
@@ -2,7 +2,10 @@ from modelcluster.models import ClusterableModel
 from wagtail.admin.edit_handlers import StreamFieldPanel
 from wagtail.contrib.settings.models import BaseSetting, register_setting
 from wagtail.core import blocks
+from wagtail.core.blocks import StreamBlock
 from wagtail.core.fields import StreamField
+
+from bc.utils.constants import RICH_TEXT_FEATURES
 
 
 class LinkBlock(blocks.StructBlock):
@@ -15,11 +18,11 @@ class LinkBlock(blocks.StructBlock):
         template = ("patterns/molecules/navigation/blocks/menu_item.html",)
 
 
-class LinkColumnWithHeader(blocks.StructBlock):
+class ColumnWithHeader(blocks.StructBlock):
     heading = blocks.CharBlock(
         required=False, help_text="Leave blank if no header required."
     )
-    links = blocks.ListBlock(LinkBlock())
+    content = blocks.RichTextBlock(features=RICH_TEXT_FEATURES)
 
     class Meta:
         template = ("patterns/molecules/navigation/blocks/footer_column.html",)
@@ -27,6 +30,14 @@ class LinkColumnWithHeader(blocks.StructBlock):
 
 @register_setting(icon="list-ul")
 class NavigationSettings(BaseSetting, ClusterableModel):
+    footer_columns = StreamField(
+        StreamBlock(
+            [("column", ColumnWithHeader())],
+            blank=True,
+            help_text="Columns of free text above the base footer.",
+            max_num=3,
+        )
+    )
     footer_links = StreamField(
         [("link", LinkBlock())],
         blank=True,
@@ -34,5 +45,6 @@ class NavigationSettings(BaseSetting, ClusterableModel):
     )
 
     panels = [
+        StreamFieldPanel("footer_columns"),
         StreamFieldPanel("footer_links"),
     ]