+++ bc/recruitment/tests/test_views.py
@@ -42,6 +42,7 @@ class TestApplyConfigKey(TestCase):
             INTERNAL_HOSTNAME,
             EXTERNAL_HOSTNAME,
         ],
+        BASE_URL="http://localhost/",
         TALENTLINK_APPLY_CONFIG_KEY_EXTERNAL="the-external-one",
         TALENTLINK_APPLY_CONFIG_KEY_INTERNAL="the-internal-one",
     )
@@ -65,6 +66,7 @@ class TestApplyConfigKey(TestCase):
             INTERNAL_HOSTNAME,
             EXTERNAL_HOSTNAME,
         ],
+        BASE_URL="http://localhost/",
         TALENTLINK_APPLY_CONFIG_KEY_EXTERNAL="the-external-one",
         TALENTLINK_APPLY_CONFIG_KEY_INTERNAL="the-internal-one",
     )
@@ -88,6 +90,7 @@ class TestApplyConfigKey(TestCase):
             INTERNAL_HOSTNAME,
             EXTERNAL_HOSTNAME,
         ],
+        BASE_URL="http://localhost/",
     )
     def test_404_when_job_id_is_badly_formatted(self):
         """This is mainly to ensure that IDs which previously raised a 500 don't."""
@@ -114,6 +117,7 @@ class TestApplyConfigKey(TestCase):
             INTERNAL_HOSTNAME,
             EXTERNAL_HOSTNAME,
         ],
+        BASE_URL="http://localhost/",
     )
     def test_view_when_job_id_does_not_exist(self):
         for job_id in [
@@ -137,6 +141,7 @@ class TestApplyConfigKey(TestCase):
             INTERNAL_HOSTNAME,
             EXTERNAL_HOSTNAME,
         ],
+        BASE_URL="http://localhost/",
     )
     def test_sidebar_normally_shown(self):
         job = TalentLinkJobFactory.create(
@@ -156,6 +161,7 @@ class TestApplyConfigKey(TestCase):
             INTERNAL_HOSTNAME,
             EXTERNAL_HOSTNAME,
         ],
+        BASE_URL="http://localhost/",
     )
     def test_sidebar_not_shown_when_job_id_does_not_exist(self):
         resp = self.client.get(